requirements.txt
appdirs==1.4.3
Django==1.10
django-debug-toolbar==1.7
django-openid-auth==0.13
gdata==2.0.18
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
PyJWT==1.4.2
pyparsing==2.2.0
python-openid==2.2.5
python-social-auth==0.3.6
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-core==1.2.0
sqlparse==0.2.3
